requirements.txt
jinja2
urllib3
cairosvg
Pillow
